Split cuelogging into separate files/modules
